# packages/openhands-agent-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenHands Agent Server",
    "version": "0.1.0",
    "description": "OpenHands Agent Server - REST/WebSocket interface for OpenHands AI Agent"
  },
  "paths": {
...
    "/api/profiles": {
      "get": {
        "tags": [
          "Profiles"
        ],
        "summary": "List LLM profiles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "profiles": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "profileId": {
                            "type": "string",
                            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"
                          },
                          "providerId": {
                            "type": "string",
                            "minLength": 1,
                            "pattern": "^[A-Za-z0-9._-]+$"
                          },
                          "model": {
                            "type": "string",
                            "minLength": 1
                          },
                          "baseUrl": {
                            "default": null,
                            "anyOf": [
                              {
                                "type": "string",
                                "format": "uri"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "openAiApiMode": {
                            "default": "chat_completions",
                            "anyOf": [
                              {
                                "type": "string",
                                "const": "chat_completions"
                              },
                              {
                                "type": "string",
                                "const": "responses"
                              }
                            ]
                          },
                          "temperature": {
                            "default": null,
                            "anyOf": [
                              {
                                "type": "number",
                                "minimum": 0
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "topP": {
                            "default": null,
                            "anyOf": [
                              {
                                "type": "number",
                                "minimum": 0,
                                "maximum": 1
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "topK": {
                            "default": null,
                            "anyOf": [
                              {
                                "type": "integer",
                                "exclusiveMinimum": 0,
                                "maximum": 9007199254740991
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "maxInputTokens": {
                            "default": null,
                            "anyOf": [
                              {
                                "type": "integer",
                                "exclusiveMinimum": 0,
                                "maximum": 9007199254740991
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "maxOutputTokens": {
                            "default": null,
                            "anyOf": [
                              {
                                "type": "integer",
                                "exclusiveMinimum": 0,
                                "maximum": 9007199254740991
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "timeoutSeconds": {
                            "default": null,
                            "anyOf": [
                              {
                                "type": "number",
                                "exclusiveMinimum": 0
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "reasoningEffort": {
                            "default": null,
                            "anyOf": [
                              {
                                "anyOf": [
                                  {
                                    "type": "string",
                                    "const": "low"
                                  },
                                  {
                                    "type": "string",
                                    "const": "medium"
                                  },
                                  {
                                    "type": "string",
                                    "const": "high"
                                  }
                                ]
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "reasoningSummary": {
                            "default": null,
                            "anyOf": [
                              {
                                "anyOf": [
                                  {
                                    "type": "string",
                                    "const": "auto"
                                  },
                                  {
                                    "type": "string",
                                    "const": "concise"
                                  },
                                  {
                                    "type": "string",
                                    "const": "detailed"
                                  }
                                ]
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "headers": {
                            "default": {},
                            "type": "object",
                            "propertyNames": {
                              "type": "string"
                            },
                            "additionalProperties": {
                              "type": "string"
                            }
                          },
                          "useProfileKeyOverride": {
                            "default": false,
                            "type": "boolean"
                          }
                        },
                        "required": [
                          "profileId",
                          "providerId",
                          "model",
                          "baseUrl",
                          "openAiApiMode",
                          "temperature",
                          "topP",
                          "topK",
                          "maxInputTokens",
                          "maxOutputTokens",
                          "timeoutSeconds",
                          "reasoningEffort",
                          "reasoningSummary",
                          "headers",
                          "useProfileKeyOverride"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "active_profile_id": {
                      "default": null,
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "profiles",
                    "active_profile_id"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Profiles"
        ],
        "summary": "Create or update LLM profile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "profileId": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"
                  },
                  "providerId": {
                    "type": "string",
                    "minLength": 1,
                    "pattern": "^[A-Za-z0-9._-]+$"
                  },
                  "model": {
                    "type": "string",
                    "minLength": 1
                  },
                  "baseUrl": {
                    "default": null,
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "uri"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "openAiApiMode": {
                    "default": "chat_completions",
                    "anyOf": [
                      {
                        "type": "string",
                        "const": "chat_completions"
                      },
                      {
                        "type": "string",
                        "const": "responses"
                      }
                    ]
                  },
                  "temperature": {
                    "default": null,
                    "anyOf": [
                      {
                        "type": "number",
                        "minimum": 0
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "topP": {
                    "default": null,
                    "anyOf": [
                      {
                        "type": "number",
                        "minimum": 0,
                        "maximum": 1
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "topK": {
                    "default": null,
                    "anyOf": [
                      {
                        "type": "integer",
                        "exclusiveMinimum": 0,
                        "maximum": 9007199254740991
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "maxInputTokens": {
                    "default": null,
                    "anyOf": [
                      {
                        "type": "integer",
                        "exclusiveMinimum": 0,
                        "maximum": 9007199254740991
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "maxOutputTokens": {
                    "default": null,
                    "anyOf": [
                      {
                        "type": "integer",
                        "exclusiveMinimum": 0,
                        "maximum": 9007199254740991
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "timeoutSeconds": {
                    "default": null,
                    "anyOf": [
                      {
                        "type": "number",
                        "exclusiveMinimum": 0
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "reasoningEffort": {
                    "default": null,
                    "anyOf": [
                      {
                        "anyOf": [
                          {
                            "type": "string",
                            "const": "low"
                          },
                          {
                            "type": "string",
                            "const": "medium"
                          },
                          {
                            "type": "string",
                            "const": "high"
                          }
                        ]
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "reasoningSummary": {
                    "default": null,
                    "anyOf": [
                      {
                        "anyOf": [
                          {
                            "type": "string",
                            "const": "auto"
                          },
                          {
                            "type": "string",
                            "const": "concise"
                          },
                          {
                            "type": "string",
                            "const": "detailed"
                          }
                        ]
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "headers": {
                    "default": {},
                    "type": "object",
                    "propertyNames": {
                      "type": "string"
                    },
                    "additionalProperties": {
                      "type": "string"
                    }
                  },
                  "useProfileKeyOverride": {
                    "default": false,
                    "type": "boolean"
                  }
                },
                "required": [
                  "profileId",
                  "providerId",
                  "model",
                  "baseUrl",
                  "openAiApiMode",
                  "temperature",
                  "topP",
                  "topK",
                  "maxInputTokens",
                  "maxOutputTokens",
                  "timeoutSeconds",
                  "reasoningEffort",
                  "reasoningSummary",
                  "headers",
                  "useProfileKeyOverride"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "profileId": {
                      "type": "string",
                      "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"
                    },
                    "providerId": {
                      "type": "string",
                      "minLength": 1,
                      "pattern": "^[A-Za-z0-9._-]+$"
                    },
                    "model": {
                      "type": "string",
                      "minLength": 1
                    },
                    "baseUrl": {
                      "default": null,
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "uri"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "openAiApiMode": {
                      "default": "chat_completions",
                      "anyOf": [
                        {
                          "type": "string",
                          "const": "chat_completions"
                        },
                        {
                          "type": "string",
                          "const": "responses"
                        }
                      ]
                    },
                    "temperature": {
                      "default": null,
                      "anyOf": [
                        {
                          "type": "number",
                          "minimum": 0
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "topP": {
                      "default": null,
                      "anyOf": [
                        {
                          "type": "number",
                          "minimum": 0,
                          "maximum": 1
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "topK": {
                      "default": null,
                      "anyOf": [
                        {
                          "type": "integer",
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "maxInputTokens": {
                      "default": null,
                      "anyOf": [
                        {
                          "type": "integer",
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "maxOutputTokens": {
                      "default": null,
                      "anyOf": [
                        {
                          "type": "integer",
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "timeoutSeconds": {
                      "default": null,
                      "anyOf": [
                        {
                          "type": "number",
                          "exclusiveMinimum": 0
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "reasoningEffort": {
                      "default": null,
                      "anyOf": [
                        {
                          "anyOf": [
                            {
                              "type": "string",
                              "const": "low"
                            },
                            {
                              "type": "string",
                              "const": "medium"
                            },
                            {
                              "type": "string",
                              "const": "high"
                            }
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "reasoningSummary": {
                      "default": null,
                      "anyOf": [
                        {
                          "anyOf": [
                            {
                              "type": "string",
                              "const": "auto"
                            },
                            {
                              "type": "string",
                              "const": "concise"
                            },
                            {
                              "type": "string",
                              "const": "detailed"
                            }
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "headers": {
                      "default": {},
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "useProfileKeyOverride": {
                      "default": false,
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "profileId",
                    "providerId",
                    "model",
                    "baseUrl",
                    "openAiApiMode",
                    "temperature",
                    "topP",
                    "topK",
                    "maxInputTokens",
                    "maxOutputTokens",
                    "timeoutSeconds",
                    "reasoningEffort",
                    "reasoningSummary",
                    "headers",
                    "useProfileKeyOverride"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Error"
          }
        }
      }
    },
...
    "/api/profiles/{name}/rename": {
      "post": {
        "tags": [
          "Profiles"
        ],
        "summary": "Rename LLM profile",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "new_name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  }
                },
                "required": [
                  "new_name"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "name",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Error"
          },
          "409": {
            "description": "Error"
          }
        }
      }
    },
    "/api/profiles/{name}/activate": {
      "post": {
        "tags": [
          "Profiles"
        ],
        "summary": "Activate LLM profile",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Error"
          }
        }
      }
    },
...
  }
}